requirements.txt
Flask~=2.1.2
numpy~=1.23.0
pandas~=1.4.3
Pillow~=9.1.1
gunicorn~=20.1.0